+++ test/transform/resource/after-delombok/GetterOnMethodErrors.java
@@ -1,9 +1,0 @@
-class PlaceFillerToMakeSurePositionIsRelevant {
-}
-class GetterOnMethodErrors {
-	private int test;
-	@java.lang.SuppressWarnings("all")
-	public int getTest() {
-		return this.test;
-	}
-}